Friday Work
Moved the book header and source files to the src folder. Added
functionality to main and interface.cpp
