+++ src/Interface.h
@@ -8,8 +8,11 @@ class InvInterface{
 		// The inventory of our books
 		Inventory books;
 		
-		enum MenuOption {}
-		const int numMenuOptions = 5;
+		enum MenuOption {View, Sell, Order, Exit};
+		
+		// Set numMenuOptions equal to the last value in the MenuOption enum
+		const int numMenuOptions = Exit;
+		
 	public:
 		// Constructors
 		InvInterface();
@@ -20,6 +23,14 @@ class InvInterface{
 		//
 		void Run();
 		int GetMenuChoice();
+		
+		string GetTitle();
+		
+		// User actions
+		void ViewInventory();
+		void SellBook(string title);
+		void OrderBook(string title);
+		
 };
 
 #endif
