+++ src/main.cpp
@@ -5,8 +5,15 @@ using namespace std;
 
 void main(){
 	
+	// Initailize
 	InvInterface program = new InvInterface();
+	
+	// Run 
 	program.Run();
 	
+	// Clearn up
+	delete program;
+	
+	// Exit
 	return;
 }
